+++ modules/firefox.nix
@@ -7,5 +7,11 @@
         "browser.tabs.closeWindowWithLastTab" = false;
       };
     };
+
+    programs.zsh.profileExtra = ''
+      if [[ $XDG_SESSION_TYPE = "wayland" ]]; then
+        export MOZ_ENABLE_WAYLAND=1
+      fi
+    '';
   };
 }